three-way merged

+++ anchongnet/app/Http/Controllers/admin/goodController.php
@@ -375,6 +375,10 @@ class goodController extends Controller
             if (!$data || Gate::denies('shopres',$data)) {
                 return null;
             }
+            //促销中不可删除
+            if ($data[0]->promotion_price > 0 ) {
+                return '促销中的商品不可删除';
+            }
             DB::beginTransaction();
             $res['spe'] = DB::table('anchong_goods_specifications')->where('gid',$aid)->delete();
             $res['stock'] = DB::table('anchong_goods_stock')->where('gid',$aid)->delete();
